requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==21.2.0
h5py==3.15.1
idna==3.11
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.8
ml-dtypes==0.2.0
numpy==1.24.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
pandas==2.1.4
Pillow==10.0.1
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pymongo==4.6.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
google-generativeai==0.8.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.3.1
scipy==1.15.3
six==1.17.0
sounddevice==0.5.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
langchain==0.0.340
openai==0.28.0
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0
sentence-transformers==2.2.2